backend/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
flake8==7.1.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
